+++ src/service/query-page-state.service.ts
@@ -8,8 +8,8 @@ import { inject, Injectable } from "@angular/core";
 import { FormControl } from "@angular/forms";
 import { BehaviorSubject, combineLatest, map, Observable, pairwise, shareReplay, startWith, switchMap } from "rxjs";
 import { DriverAction } from "../concept/action";
-import {createSigmaRenderer, GraphVisualiser} from "../framework/graph-visualiser";
-import { defaultSigmaSettings } from "../framework/graph-visualiser/defaults";
+import { GraphVisualiser } from "../framework/graph-visualiser";
+import { createSigmaRenderer, defaultSigmaSettings } from "../framework/graph-visualiser/sigma-settings";
 import { newVisualGraph, VisualGraph } from "../framework/graph-visualiser/graph";
 import { Layouts } from "../framework/graph-visualiser/layouts";
 import { detectOS } from "../framework/util/os";
